+++ src/target/excel/compilation_target.rs
@@ -1,7 +1,7 @@
 //!
 use super::super::{file_backer_upper, CompilationTarget};
 use super::Excel;
-use crate::{Result, Template};
+use crate::{Module, Result};
 use umya_spreadsheet as u;
 
 impl CompilationTarget for Excel<'_> {
@@ -10,7 +10,7 @@ impl CompilationTarget for Excel<'_> {
         Ok(())
     }
 
-    fn write(&self, template: &Template) -> Result<()> {
+    fn write(&self, module: &Module) -> Result<()> {
         let mut spreadsheet = self.open_spreadsheet()?;
 
         self.create_worksheet(&mut spreadsheet)?;
@@ -19,7 +19,7 @@ impl CompilationTarget for Excel<'_> {
         // to get it but I couldn't get the mutable references to work out.
         let worksheet = self.get_worksheet_mut(&mut spreadsheet)?;
 
-        self.build_worksheet(template, worksheet)?;
+        self.build_worksheet(module, worksheet)?;
 
         u::writer::xlsx::write(&spreadsheet, self.path.clone()).map_err(|e| {
             self.runtime.output_error(format!(
@@ -42,8 +42,8 @@ mod tests {
         let test_file = &TestSourceCode::new("xlsx", "foo,bar,baz");
         let runtime = test_file.into();
         let target = Excel::new(&runtime, test_file.output_file.clone());
-        let template = Template::compile(&runtime).unwrap();
+        let module = Module::compile(&runtime).unwrap();
 
-        assert!(target.write(&template).is_ok());
+        assert!(target.write(&module).is_ok());
     }
 }