+++ tests/all_features.rs
@@ -1,31 +1,31 @@
-use csvpp::Template;
+use csvpp::Module;
 mod common;
 
 #[test]
 fn all_features_csv() {
     let s = common::Setup::from_fixture("all_features", "csv");
-    let template = Template::compile(&s.runtime).unwrap();
+    let module = Module::compile(&s.runtime).unwrap();
     let target = s.runtime.target().unwrap();
 
-    assert!(target.write(&template).is_ok());
+    assert!(target.write(&module).is_ok());
 }
 
 #[test]
 fn all_features_shorthand_csv_no_code_section() {
     let s = common::Setup::from_fixture("all_features_shorthand", "csv");
-    let template = Template::compile(&s.runtime).unwrap();
+    let module = Module::compile(&s.runtime).unwrap();
     let target = s.runtime.target().unwrap();
 
-    assert!(target.write(&template).is_ok());
+    assert!(target.write(&module).is_ok());
 }
 
 #[test]
 fn all_features_excel() {
     let s = common::Setup::from_fixture("all_features", "xlsx");
-    let template = Template::compile(&s.runtime).unwrap();
+    let module = Module::compile(&s.runtime).unwrap();
     let target = s.runtime.target().unwrap();
 
-    assert!(target.write(&template).is_ok());
+    assert!(target.write(&module).is_ok());
 }
 
 // TODO: